feat: selector pagination hook처리
